backend/requirements.txt
fastapi
uvicorn
sqlglot
python-multipart
python-dotenv
g4f
sqlalchemy
psycopg2-binary
passlib[bcrypt]
python-jose[cryptography]